split out cli io helpers into copybook-cli-io microcrate
