+++ Cargo.toml
@@ -13,6 +13,7 @@ members = [
     "copybook-codepage",
     "copybook-charset",
     "copybook-options",
+    "copybook-cli-io",
     "copybook-core",
     "copybook-dialect",
     "copybook-codec",
@@ -63,6 +64,7 @@ copybook-determinism = { version = "=0.4.3", path = "copybook-determinism" }
 copybook-codepage = { version = "=0.4.3", path = "copybook-codepage" }
 copybook-charset = { version = "=0.4.3", path = "copybook-charset" }
 copybook-options = { version = "=0.4.3", path = "copybook-options" }
+copybook-cli-io = { version = "=0.4.3", path = "copybook-cli-io" }
 copybook-core = { version = "=0.4.3", path = "copybook-core" }
 copybook-dialect = { version = "=0.4.3", path = "copybook-dialect" }
 copybook-codec = { version = "=0.4.3", path = "copybook-codec" }
