+++ copybook-cli-io/Cargo.toml
@@ -1,0 +1,24 @@
+# SPDX-License-Identifier: AGPL-3.0-or-later
+[package]
+name = "copybook-cli-io"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Focused CLI I/O helpers for atomic output writes and stdin-aware text loading."
+readme = "README.md"
+keywords = ["cli", "io", "atomic", "copybook"]
+categories = ["command-line-utilities", "filesystem"]
+
+[dependencies]
+tempfile.workspace = true
+tracing.workspace = true
+
+[dev-dependencies]
+anyhow.workspace = true
+
+[lints]
+workspace = true
