+++ copybook-cli/src/commands/encode.rs
@@ -3,11 +3,12 @@
 
 use crate::exit_codes::ExitCode;
 use crate::utils::{
-    ParseOptionsConfig, apply_field_projection, atomic_write, build_parse_options,
-    determine_exit_code, read_file_or_stdin,
+    ParseOptionsConfig, apply_field_projection, build_parse_options, determine_exit_code,
+    read_file_or_stdin,
 };
 use crate::{write_stderr_all, write_stdout_all};
 use anyhow::{anyhow, bail};
+use copybook_cli_io::atomic_write;
 use copybook_codec::{Codepage, EncodeOptions, FloatFormat, RecordFormat};
 use copybook_core::parse_copybook_with_options;
 use std::fmt::Write as _;
